LiveGraph improvements
Display 3 small graphs instead of one big one
