+++ src/StaticConfiguration.ts
@@ -59,15 +59,15 @@ class StaticConfiguration {
   };
 
   // Line colors are picked in the order of this array.
-  public static readonly liveGraphColors = [
-    '#f9808e',
-    '#80f98e',
-    '#808ef9',
-    '#eb8634',
-    '#4cfaee',
-    '#fa4cfa',
-    '#6b32a8',
-  ];
+  public static readonly liveGraphColors: { [key: string]: string } = {
+    accx: '#f9808e',
+    accy: '#80f98e',
+    accz: '#808ef9',
+    magx: '#eb8634',
+    magy: '#4cfaee',
+    magz: '#fa4cfa',
+    l: '#6b32a8',
+  };
 
   // What will the min and max y-values on the livegraph be?
   public static readonly liveGraphValueBounds = {
