+++ src/script/livedata/CombinedData.ts
@@ -58,12 +58,12 @@ class CombinedLiveData implements LiveData<FlatCombinedData> {
       [accelerometerData, magnetometerData, lightData],
       ([a, m, l]) => {
         const data = {
-          accx: a.x,
-          accy: a.y,
-          accz: a.z,
-          magx: m.x,
-          magy: m.y,
-          magz: m.z,
+          accx: a.accx,
+          accy: a.accy,
+          accz: a.accz,
+          magx: m.magx,
+          magy: m.magy,
+          magz: m.magz,
           light: l.l,
         };
         this.dataBuffer.addValue(data);
